backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
gunicorn>=21.2,<22.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
whitenoise>=6.7,<7.0
paho-mqtt>=1.6,<2.0
